recapp-api/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2019.11.28
cycler==0.10.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
isort==4.3.21
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
matplotlib==3.0.0
mccabe==0.6.1
numpy==1.15.3
pandas==0.24.2
Pillow==7.1.2
pylint==2.4.4
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2019.1
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
